requirements.txt
accelerate==1.2.1
bitsandbytes==0.44.0; platform_system != "Windows"
datasets==2.19.0
dvc==3.53.1
evaluate==0.4.2
huggingface-hub==1.0.0rc6
langdetect==1.0.9
numpy==1.26.4
pandas==2.2.2
peft==0.12.0
pyarrow==15.0.2
PyYAML==6.0.2
pytest==8.2.1
rich==13.7.1
scikit-learn==1.4.2
sentencepiece==0.1.99
tensorboard==2.16.2
torch==2.2.2
transformers @ git+https://github.com/huggingface/transformers.git
trl==0.11.4
tqdm==4.66.4
typer[all]==0.12.3
click==8.1.7
requests==2.32.3
python-dotenv==1.0.1
